CHG: split out some age and length specific reports CHG: length bins in the @model must be integers (used to be Double), makes it analogous to age bins FIX: TestModel -> needed the first block to be @model

+++ CASAL2/source/Length/Processes/Children/RecruitmentConstant.h
@@ -50,7 +50,7 @@ private:
   vector<Double>              proportions_;
   OrderedMap<string, Double>  proportions_categories_;
   Double                      r0_;
-  vector<Double>              length_bins_;
+  vector<unsigned>            length_bins_;
   accessor::Categories        partition_;
   Double                      r0_by_length_bin_;
 };
